add Process specific methods code refactored

+++ src/processor.cpp
@@ -4,10 +4,10 @@
 float Processor::Utilization() {
     std::vector<float> cpu = LinuxParser::CpuUtilization();
 
-    float user = static_cast<float>(cpu[0]);
-    float nice = static_cast<float>(cpu[1]);
-    float system = static_cast<float>(cpu[2]);
-    float idle = static_cast<float>(cpu[3]);
+    float user = cpu[0];
+    float nice = cpu[1];
+    float system = cpu[2];
+    float idle = cpu[3];
 
     float total = user + nice + system + idle;
     float busy = total - idle;
